economy/requirements.txt
click==7.1.1
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Frozen-Flask==0.15
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
SQLAlchemy==1.3.15
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.3.1
